+++ Changes.md
@@ -34,6 +34,11 @@ Build
 
 - Cortex : Updated to version 10.4.7.0.
 
+Documentation
+-------------
+
+- Updated the "Contact Sheet Generation" example to use Cycles instead of Appleseed.
+
 1.2.2.0 (relative to 1.2.1.1)
 =======